refactor(ui): flatten `sensors/` dir
- mostly same reasons as previous commits
- `sensor-details` and `sensor-list` had one file per folder, which is redundant
- notably, `utils.ts` is not a component and was used outside of `sensors` too
