+++ ui/src/app/sensors/index.ts
@@ -1,4 +1,4 @@
-import {SensorsContainer} from './components/sensors-container';
+import {SensorsContainer} from './sensors-container';
 
 export default {
     component: SensorsContainer
